pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "psycopg2-binary>=2.9.11",
    "sentry-sdk>=2.52.0",
    "sqlmodel>=0.0.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "python-dotenv>=1.2.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
